+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -7,3 +7,4 @@
 ### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.12.0
 
 - Updated `Microsoft.Azure.WebJobs.Extensions.CosmosDB` reference to 4.9.0
+- Return a successful result with a null value when a Cosmos document cannot be found (#2942/#2545)